mno-async_writes/client: Added base class TestRPCServer and first, incomplete implementation TestRPCServerDIR.
In order to write unit tests which allow to test the interaction with XtreemFS servers, it was necessary to implement dummy XtreemFS servers in C++. The new class TestRPCServer, which uses synchronous boost::asio functions, provides the base for this. Minimal test server can be implemented now to test the correct functionality of the C++ client.

git-svn-id: https://xtreemfs.googlecode.com/svn/branches/mno-async_writes@2970 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534