client: Refactored SSL passphrase command line parsing code.
implicit_value() is no longer used and therefore its limitations do no longer apply. Since revision 2911 it was not possible to specify arguments with a " " instead of a "=" after the option e.g., --pkcs12-file-path certfile.p12.

This functionality was restored. If the user wants to let the client the passphrase from stdin, now he has to provide "-" as passphrase or omit the passphrase option.

git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3007 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534