foundation: Modified interface of LifeCycleThread.notifyCrashed().
The notifyCrashed(Exception) function is used by threads to notify the process of a crash of a single thread. Consequently, a registered listener can crash the whole process and report the error to the user.

However, the current interface only caught "Exception"s and no "Error"s e.g., an AssertionError. This commit changes the interface of notifyCrashed() to use the more general type "Throwable" as input parameter. Further commits will follow which will update the correct usage of notifyCrashed() accordingly.

git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3149 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534