server: Record stack traces of responseAvailable() calls in case of stripe operations.
In case of a striped read, the OSD may hang at the get() of a RPCResponse. However, this mustn't be possible since waitForResponses() ensured beforehand that responses.length() executions of responseAvailable() were seen. In that case responseAvailable() was probably called multiple times for the same RPCResponse. This change will detect this situation and throw a RuntimeException.

git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3215 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534