server: DIR: Do no longer swallow a BabuDBException in registerOperations().
This lead to a situation where the startup of the DIR took too long: the replicated DIR database was not able to contact the other replicas and therefore the registration of each of the 10 database-related operations was blocked for several minutes. Now the startup will be immediately aborted after the first error.

git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3502 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534