client: Refactored UserMapping design.
So far, UserMappingUnix and UserMappingGridmap* were the same. Now, there are two different classes:
- SystemUserMapping
- UserMapping (e.g. GridmapGlobus)

A UserMapping allows to transform between local and global names e.g., transform the DN of a certificate to the login name "mberlin".

A SystemUserMapping only has the call "GetUserCredentialsForCurrentUser" now which fills a UserCredentials object with the username and groups of the current user. If an additional UserMapping is registered at the SystemUserMapping, it will also use that UserMapping to transform the names e.g., the user "mberlin" will be transformed to his respective DN entry.

This commit breaks the FuseAdapter. Later it will be adapted to this change.

git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3581 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534