server: OSD: Fixed bug in Read-Only Replication. Under high load, read requests, which trigger the on-demand replication of an object, could get lost and not replied by the OSD. In that case, the client had to retry and the access on a partial replica took longer than usual.
This was caused by internally overwriting existing metadata for a pending request. Requests are added to a map "waitingRequests" first and later moved to the map "objectsInProgress". In rare cases, the move does not happen at the same time as the request will be enqueued. In these cases, the existing request (for the same object!) in "waitingRequests" was overwritten and never processed.