Introduce ConnectionData datastructure and Controller class.
Introduce ConnectionData datastructure and refactor all the other classes to use it.
Introduce controller class and move the scheduleXYBenchmark-methods this class.
Rename scheduleXYBenchmark-methods to startXYBenchmark-methods.
Introduce new startBenchmarks-method to reduce code duplication in the
startXY-Benchmark-Methods.
Introduce new client handling (managed in controller). Move the scrub() and
deleteVolumeIfExisting methods to controller.