servers: views: Recover invalidated replicas and locked xLocSets after a MRC crashed.
The lock is stored in the BabuDB database and contains information about the
MRC instance which locked the xLocSet.
If the xLocSet is locked by the current instance, additional operations
changing the xLocSets are forbidden.
If the xLocSet was locked by a previous instance, it is assumed a crash occured
and the xLocSet has to be released and the replicas revalidated.