tests: views: Fix reading with an invalid view and redirection errors.
Since the syncCall is handling view errors transparent, the test did no
longer get any invalid view exceptions. To ensure invalid view errors
are passed on, the maxRetries option has to be set to 1, because invalid
view exceptions are only thrown if no retry is left.

Redirection errors occured, when the old primary was removed and backups
still tried to redirect to it until the lease timed out. To prevent
this, the second replica, that is probably a backup, will be removed.