client: Implemented multi-homing support in C++ libxtreemfs for Linux.
If multiple addresses for the same UUID are available, the first address
which matches a locally available network will be preferred.

In helper.cpp a new method "GetNetworks()" was introduced which
enumerates the available network interfaces.

Added corresponding unit test helper_test.cpp for the new method.

Added unit test client_implementation_test.cpp which tests the correct
functioning of the multi-homing support.