servers: cpp: Clarifiy GetSuitableOSDs interface documentation.
The numberOfOsds paramater is only relevant for grouping policies and will be
ignored by filtering and sorting policies. If a particular number of OSDs is
required by the caller, the returned list has to be reduced by the caller.

This lead to an erorr in the FileScrubber, where every suitable OSD was
assigned to new replicas.