servers: views: Return LeaseState to speed up invalidation.
Before the invalidation method had to wait until
1) a majority did respond and
2) either the primary or all replicas did respond
   or in case both could not be satisfied, until the lease did time out.

This did result in long delays if either 1) no primary was elected at the time
of invalidation and at least one replica was unreachable, or 2) the primary
itself did not respond.

With this commit an additional LeaseState is returned with invalidation
responses, which allows to determine if a primary did exist at all and fixes
delays as described in the first case.